requirements.txt
sqlalchemy==2.0.35
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
fastapi==0.110.0
aiosqlite==0.20.0
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
isort==5.13.2
black==24.8.0